requirements.txt
Django>=4.2
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
pdfplumber>=0.11.0
openpyxl>=3.1.2
python-docx>=1.0.0
